src/python/pyproject.toml
[tool.poetry]
name = "imdb-graphql"
version = "0.1.0"
description = "GraphQL API for IMDB"
authors = ["kiendang <kiend@pm.me>"]

[tool.poetry.dependencies]
python = "^3.6"
sqlalchemy = "^1.2"
graphene = "^2.1"
psycopg2-binary = "^2.7"
flask = "^1.0"
flask-graphql = "^2.0"
graphene-sqlalchemy = "^2.1"
sqlalchemy_opentracing = "^0.1.6"
jaeger-client = "3.13.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
ipython = "^6.5"
python-language-server = "^0.21.2"
